requirements.txt
scapy
netifaces
python-nmap
pycairo
PyGObject
fabric
matplotlib
